Persisted configuration between ppx invocations
This solves the long-standing problems of call sites not being wrapped
and the default annotation not being visible across modules. These are
just two instances of the more general problem of ppx invocations
working on independent compilation units.

Configuration is shared across compilation units by writing to a file.
The default annotation has also been augmented to reset configuration --
assuming that the module it is declared in comes first in a topological
ordering (i.e. other modules open it), this prevents problems arising
from stale configuration.