Persistent configuration is no longer reset
The default annotation is not a reliable way to reset it, as there is no
guarantee that the file it is in will be processed during compilation.

The solution for now is to remove this side effect from the default
annotation, and to let the state become stale, so the features of the
library do not fail due to incremental recompilation. Between having
stale configuration and features of the library failing randomly, the
former seems like the lesser evil, considering we need to rebuild to fix
it either way...