Manually install before testing
INSTALL defaults to true, which does an `opam install` followed by an
`opam remove`. The last bit causes the examples to fail to build. There
doesn't seem to be an option to not remove the package after
installation, so we don't disable INSTALL entirely and manually install
before testing the examples.